requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
numpy<2.0.0,>=1.24.0
scipy==1.14.1
scikit-learn==1.5.2
pandas==2.2.3
openpyxl==3.1.2
sentence-transformers==3.3.1
torch==2.5.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
python-multipart==0.0.17
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
passlib==1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0

